+++ lib/src/models/send_message_configuration.dart
@@ -19,8 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import 'package:audio_waveforms/audio_waveforms.dart';
-import 'package:chatview/src/values/enumaration.dart';
+import 'package:chatview/chatview.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:image_picker/image_picker.dart';
@@ -71,6 +70,9 @@ class SendMessageConfiguration {
   /// Styling configuration for recorder widget.
   final VoiceRecordingConfiguration? voiceRecordingConfiguration;
 
+  /// Configuration for cancel voice recording
+  final CancelRecordConfiguration? cancelRecordConfiguration;
+
   const SendMessageConfiguration({
     this.textFieldConfig,
     this.textFieldBackgroundColor,
@@ -87,6 +89,7 @@ class SendMessageConfiguration {
     this.enableGalleryImagePicker = true,
     this.voiceRecordingConfiguration,
     this.micIconColor,
+    this.cancelRecordConfiguration,
   });
 }
 
@@ -237,4 +240,22 @@ class VoiceRecordingConfiguration {
 
   /// Applies color to mic and stop icon.
   final Color? recorderIconColor;
+}
+
+class CancelRecordConfiguration {
+  /// Configuration for cancel voice recording
+  const CancelRecordConfiguration({
+    this.cancelRecordIcon,
+    this.cancelRecordIconColor,
+    this.onCancelVoiceRecord,
+  });
+
+  /// An icon for cancelling voice recording.
+  final Widget? cancelRecordIcon;
+
+  /// Cancel record icon color
+  final Color? cancelRecordIconColor;
+
+  /// Provides callback on voice record cancel
+  final VoidCallBack? onCancelVoiceRecord;
 }